# src/main/java/com/darmokhval/CarManagementService/model/dto/AveragePriceDTO.java
package com.darmokhval.CarManagementService.model.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Schema
public class AveragePriceDTO {
    private String location;
    private Double averagePrice;
}