# src/main/resources/specification/advertisement-management-service.openapi.yml
openapi: 3.0.0
info:
  title: Car Management Service API
  description: API documentation for the Car Management Service.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/advertisements:
    post:
      summary: Create Advertisement
      description: Creates a new advertisement.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                advertisement:
                  $ref: '#/components/schemas/AdvertisementDTO'
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Advertisement created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisementDTO'
    get:
      summary: Get Advertisements
      description: Retrieves a list of advertisements.
      parameters:
        - name: location
          in: query
          description: Filter advertisements by location
          schema:
            type: string
      responses:
        '200':
          description: List of advertisements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdvertisementDTO'
  /api/advertisements/{id}:
    get:
      summary: Get Advertisement by ID
      description: Retrieves an advertisement by its ID.
      parameters:
        - name: id
          in: path
          description: ID of the advertisement to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Advertisement retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisementDTO'
    put:
      summary: Update Advertisement by ID
      description: Updates an existing advertisement.
      parameters:
        - name: id
          in: path
          description: ID of the advertisement to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                car:
                  $ref: '#/components/schemas/AdvertisementDTO'
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Advertisement updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertisementDTO'
    delete:
      summary: Delete Advertisement by ID
      description: Deletes an advertisement by its ID.
      parameters:
        - name: id
          in: path
          description: ID of the advertisement to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Advertisement deleted successfully
          content:
            application/json:
              schema:
                type: string
  /api/advertisements/details:
    get:
      summary: Get Car Details
      description: Retrieves details of cars for authorized users.
      responses:
        '200':
          description: List of car details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDetailsDTO'
  /api/advertisements/me:
    get:
      summary: Find My Advertisements
      description: Retrieves advertisements posted by the current user.
      responses:
        '200':
          description: List of advertisements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdvertisementDTO'
  /api/advertisements/average-price:
    get:
      summary: Get Average Price
      description: Retrieves the average price of advertisements.
      parameters:
        - name: location
          in: query
          description: Filter average price by location
          schema:
            type: string
      responses:
        '200':
          description: Average price retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AveragePriceDTO'
  /api/advertisements/{id}/views:
    get:
      summary: Get Views for Time Period
      description: Retrieves the number of views for an advertisement in a given time period.
      parameters:
        - name: id
          in: path
          description: ID of the advertisement
          required: true
          schema:
            type: integer
            format: int64
        - name: period
          in: query
          description: Time period (e.g., "day", "week", "month")
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Number of views retrieved successfully
          content:
            application/json:
              schema:
                type: integer
components:
  schemas:
    AdvertisementDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        model:
          type: string
        brand:
          type: string
        description:
          type: string
        carPhoto:
          type: string
          format: binary
        price:
          type: integer
          format: int32
        location:
          type: string
        currencyType:
          type: string
        photoPath:
          type: string
        views:
          type: array
          items:
            $ref: '#/components/schemas/AdvertisementView'
    AveragePriceDTO:
      type: object
      properties:
        location:
          type: string
        averagePrice:
          type: number
          format: double
    CarDetailsDTO:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
    AdvertisementView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        advertisement:
          $ref: '#/components/schemas/AdvertisementDTO'
        timestamp:
          type: string
          format: date-time
        viewerId:
          type: integer
          format: int64
    ErrorDTO:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
        details:
          type: string
    InappropriateWordDTO:
      type: object
      properties:
        content:
          type: string
    LoginDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senderId:
          type: integer
          format: int64
        recipientId:
          type: integer
          format: int64
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    ResponseUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        isSeller:
          type: boolean
        isPremium:
          type: boolean